Refactor program to support library-style usage, add wasm crate
