+++ Cargo.toml
@@ -1,4 +1,5 @@
 [workspace]
 members = [
-  "crates/fta"
+  "crates/fta",
+  "crates/fta_wasm"
 ]
