+++ crates/fta/src/complexity.rs
@@ -69,6 +69,7 @@ impl Visit for ComplexityVisitor {
     }
 }
 
+#[allow(dead_code)]
 pub fn cyclomatic_complexity(module: Module) -> usize {
     let mut visitor = ComplexityVisitor::new();
     visitor.visit_module(&module);
