+++ crates/fta/src/parse_module.rs
@@ -4,6 +4,7 @@ use swc_common::SourceMap;
 use swc_ecma_ast::{EsVersion, Module};
 use swc_ecma_parser::{error::Error, lexer::Lexer, Parser, Syntax, TsConfig};
 
+#[allow(dead_code)]
 pub fn parse_module(source: &str) -> (Result<Module, Error>, usize) {
     let line_count = source.matches('\n').count() + 1;
     let cm: Lrc<SourceMap> = Default::default();
