+++ crates/fta_wasm/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "fta_wasm"
+version = "0.1.7"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+fta = { path = "../fta" }
+serde = { version = "1.0", features = ["derive"] }
+serde-wasm-bindgen = "0.3"
+wasm-bindgen = "0.2.85"
